Pull in patch from current: Fix (art): Race conditions make it hard for us to know if the interrupt was for us. Return -1 instead of 0 if we did not service the interrupt.
XXX - this is a horrible kludge, but should fix the stray interrupt problem
	that so many people have seen.

+++ sys/arch/mvme88k/ddb/db_trace.c
@@ -1,3 +1,4 @@
+/*	$OpenBSD: db_trace.c,v 1.28 1995/04/19 22:37:27 smurph Exp $	*/
 /*
  * Mach Operating System
  * Copyright (c) 1993-1991 Carnegie Mellon University
@@ -146,7 +147,7 @@ db_setf_regs(
 	*regp = *valuep;
 }
 
-#define N(s, x)  {s, (int *)&(((db_regs_t *) 0)->x), db_setf_regs}
+#define N(s, x)  {s, (long *)&(((db_regs_t *) 0)->x), db_setf_regs}
 
 struct db_variable db_regs[] = {
     N("r1", r[1]),     N("r2", r[2]), 	 N("r3", r[3]),	   N("r4", r[4]),
